+++ src/components/Header.tsx
@@ -1,6 +1,6 @@
 import { cn } from "@/lib/utils"
 import { operationStore } from "@/store/operation"
-import { Switch } from "@components/ui/switch"
+import { Switch } from "@components/shadcn/switch"
 import { Avatar, AvatarFallback, AvatarImage } from "@radix-ui/react-avatar"
 import {
   DropdownMenu,
