+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client"
 
 import Modal from "@components/Modal"
 import App from "./App"
-import { Toaster } from "./components/ui/sonner"
+import { Toaster } from "./components/shadcn/sonner"
 import "./styles/global.css"
 
 createRoot(document.getElementById("root")!).render(
